✨ [Feature] 네비게이션 기능 연결 및 shadow
네비게이션 기능 연결 및 쉐도우

+++ Widi/Widi/Common/UIComponents/Diary/SelectedImagePreview.swift
@@ -43,14 +43,16 @@ struct SelectedImagePreview: View {
                 displayImage()
             })
             
-            Button(action: {
-                onDelete()
-            }, label: {
-                Image(.closeX)
-                    .background(Color.whiteBlack.opacity(0.6))
-                    .clipShape(Circle())
-                    .padding(8)
-            })
+            if showDeleteButton {
+                Button(action: {
+                    onDelete()
+                }, label: {
+                    Image(.closeX)
+                        .background(Color.whiteBlack.opacity(0.6))
+                        .clipShape(Circle())
+                        .padding(8)
+                })
+            }
         })
     }
